+++ services/aiService.js
@@ -431,6 +431,7 @@ class AiService {
 
       let payload = { engine: engine };
       let token = jwt.sign(payload, secret);
+      console.log("token: ", token)
       axios({
         url: base_url + "/id/" + content_id + "/namespace/" + namespace_id + "/" + token,
         headers: {
@@ -486,7 +487,6 @@ class AiService {
       })
     })
   }
-
 
 }